rip_scripts/sheets/web_preview/requirements.txt
flask >= 1.0.2, < 2.0.0
pillow >= 5.0.0, < 6.0.0
requests >= 2.18.4, < 3.0.0
